experimental upgrade to 0.9 branch of buildbot
This upgrades the minimal buildbot version to 0.9.7. The configuration changes
made are based on a 'fix error in the log' approach. No deep insight in the
actual API changes was gained from it. Stuff will break.

For now the authentication is disabled but will be added in a future commit
again.