polyjit: make gitpoller aware of new codebase dict.
The new codebases dict is only used as a fallback, if we cannot fill out
the ChangeSource object with enough data from the repo. Therefore, keys
like 'branch' only allow a single string as argument (a default value).

Everything else can be configured as usual on the scheduler/git-poller side.